Use the last 2.5 hours
The README specifies that the last 8.5 hours are used, while actually it uses the last 2.5 hours from 8 hours ago (so a time frame between 10.5 hours and 8 hours ago) which means the reports will be outdated by 8 hours at the time of reporting.
This will report the events from the last 2.5 hours until now.