Add pepper environment variable
In the process, I removed option of defining a config.yaml file.
It wouldn't be secure, and I don't want to promote insecure ways of using it.
Remove option of defining a config.yaml file

It wouldn't be secure, and I don't want to promote insecure ways of using it.